chore(Profile): refacto profile-config-service and remove profile-service
