+++ src/features/discord/functions/add-discord-dev-prefix.ts
@@ -4,8 +4,9 @@ export function addDiscordDevPrefix(
   message: Readonly<string>,
   nickname: Readonly<string>
 ): string {
-  if (_.isNull(nickname) || _.isEmpty(nickname)) {
+  if (!_.isString(nickname) || _.isEmpty(nickname)) {
     return `**[dev]** ${message}`;
   }
+
   return `**[dev - ${nickname}]** ${message}`;
 }
